카페 정보 및 찜 로직을 useCafeInfo 훅으로 분리
